double -> float

+++ Solution/Wiki-CUDA-LSA/LargeMatrixParallelSVD/svd.cuh
@@ -2,16 +2,16 @@
 #ifndef _INC_SVD // include guard for 3rd party interop
 #define _INC_SVD
 
-void transpose(int rows, int cols, double * matrix);
+void transpose(int rows, int cols, float * matrix);
 
-void svd(int rows, int cols, double * matrix_A, double * matrix_U, double * array_S, double * matrix_VT);
+void svd(int rows, int cols, float * matrix_A, float * matrix_U, float * array_S, float * matrix_VT);
 
-void svd_r(int rows, int cols, double * matrix_U, double * array_S, double * matrix_VT, double * matrix_A);
+void svd_r(int rows, int cols, float * matrix_U, float * array_S, float * matrix_VT, float * matrix_A);
 
-void approximate_svd(int rows, int cols, int k, double * matrix_A, double * matrix_U, double * array_S, double * matrix_VT);
+void approximate_svd(int rows, int cols, int k, float * matrix_A, float * matrix_U, float * array_S, float * matrix_VT);
 
-void multiplyByDiagonalMatrix(int rows, int cols, double * matrix, double * array);
+void multiplyByDiagonalMatrix(int rows, int cols, float * matrix, float * array);
 
-void multiply(int rows, int k, int cols, double * matrixA, double * matrixB, double * result);
+void multiply(int rows, int k, int cols, float * matrixA, float * matrixB, float * result);
 
 #endif // _INC_SVD
