+++ Solution/Wiki-CUDA-LSA/ParallelTFIDF/tfidf.cu
@@ -3,7 +3,7 @@
 #include<cmath>
 
 //TODO: parallelize
-void tfidf(int rows, int cols, double * matrix) {
+void tfidf(int rows, int cols, float * matrix) {
 	int numDocs = rows;
 	int numTerms = cols;
 	for (int termIdx = 0; termIdx < numTerms; termIdx++) {
@@ -13,7 +13,7 @@ void tfidf(int rows, int cols, double * matrix) {
 				includeDocs++;
 			}
 		}
-		double termIdf = numDocs / (log10(includeDocs) + 1);
+		float termIdf = numDocs / (log10(includeDocs) + 1);
 		for (int docIdx = 0; docIdx < numDocs; docIdx++) {
 			matrix[docIdx * cols + termIdx] *= termIdf;
 		}
